@@ -1,8 +1,7 @@
 FactoryGirl.define do
   factory :user do
     confirmed_at Time.now
-    sequence(:name) { |n| "user#{n}" }
-    sequence(:email) { |n| "user#{n}@example.com" }
+    sequence(:email) { |n| "test_user#{n}@example.com" }
     password "please123"
 
     trait :admin do
