@@ -10,7 +10,7 @@ feature "Creating a Plan" do
     click_on 'New Plan'
     expect(page).to have_title(full_title('New Plan'))
     fill_in 'Name', with: plan_attributes[:name]
-    expect { click_on 'Create Plan' }.to change(Plan, :count).by(1)
+    expect { click_on 'Create' }.to change(Plan, :count).by(1)
     expect(user.plans.first.name).to eq(plan_attributes[:name])
     expect(page).to have_title(plan_attributes[:name])
   end
@@ -24,7 +24,7 @@ feature "Creating a Plan" do
     click_on 'New Plan'
     expect(page).to have_title(full_title('New Plan'))
     fill_in 'Name', with: plan_attributes[:name]
-    click_on 'Create Plan'
+    click_on 'Create'
     expect(page).to have_title(plan_attributes[:name])
   end
 end
