@@ -19,20 +19,38 @@ describe Course do
 
   describe "uniqeness validations" do
     before { course.save }
-
+    # This needs to go down here because of the way the 'shoulda' uniqueness
+    # matcher works.
     it { should validate_uniqueness_of(:title).case_insensitive }
+    it do
+      should validate_uniqueness_of(:level)
+        .scoped_to(:department)
+        .with_message("duplicate level for that department")
+    end
+  end
+
+  context "when a course with the same department and level already exists" do
+    before do
+      create :course, department: course.department, level: course.level
+    end
+    it "should not be valid" do
+      expect(subject).to be_invalid
+    end
   end
 
   context "with acceptable attributes" do
     it { should be_valid }
   end
 
-  context "using a bad format title" do
-    before { course.title = 'wEirD caPitalization' }
-
-    it "should fix it" do
-      course.save
-      expect(course.title).to eq("Weird Capitalization")
+  describe "#destroy" do
+    before { course.save }
+    it "should also destroy associated taken_courses" do
+      create :taken_course, course: course
+      expect { course.destroy }.to change(TakenCourse, :count).by(-1)
+    end
+    it "should also destroy associated planned_courses" do
+      create :planned_course, course: course
+      expect { course.destroy }.to change(PlannedCourse, :count).by(-1)
     end
   end
 
